Gata si incubarea

+++ Randemic_cod/RandomGen/HeaderVirus.h
@@ -2,8 +2,8 @@
 //tot felul de constante si variabile pentru virus si simulare
 
 extern int infectabilitate; //procent din 100% virus
-extern int timpRecuperare;//frameuri virus
-extern int timpIncubare;//timp incubare virus
+extern long timpRecuperare;//frameuri virus
+extern long timpIncubare;//timp incubare virus
 extern int mortalitate;//procent din 100% virus
 extern int imunitateGrup;//cat % din populatie este imuna la virus
